feat(clock): changed day format (long -> short)

+++ src/features/clock/Clock.types.ts
@@ -7,7 +7,7 @@ export const DATE_TIME_OPTIONS = {
 		second: '2-digit' as const
 	},
 	date: {
-		weekday: 'long' as const,
+		weekday: 'short' as const,
 		year: 'numeric' as const,
 		month: 'long' as const,
 		day: 'numeric' as const
